+++ framework/wazuh/configuration.py
@@ -7,7 +7,6 @@ from xml.etree.ElementTree import fromstring
 from os import listdir, path as os_path
 import re
 from wazuh.exception import WazuhException
-from wazuh.agent import Agent
 from wazuh import common
 from wazuh.utils import cut_array
 
@@ -33,6 +32,10 @@ conf_sections = {
     'open-scap': {
         'type': 'simple',
         'list_options': ['content']
+    },
+    'cis-cat': {
+        'type': 'simple',
+        'list_options': []
     },
     'rootcheck': {
         'type': 'simple',
@@ -116,7 +119,7 @@ def _read_option(section_name, opt):
                 opt_value[a] = opt.attrib[a]
             # profiles
             profiles_list = []
-            for profiles in opt.getchildren():
+            for profiles in opt.iter():
                 profiles_list.append(profiles.text)
 
             if profiles_list:
@@ -154,11 +157,11 @@ def _conf2json(src_xml, dst_json):
     Parses src_xml to json. It is inserted in dst_json.
     """
 
-    for section in src_xml.getchildren():
-        section_name = 'open-scap' if section.tag.lower() == 'wodle' else section.tag.lower()
+    for section in list(src_xml):
+        section_name = section.attrib['name'] if section.tag.lower() == 'wodle' else section.tag.lower()
         section_json = {}
 
-        for option in section.getchildren():
+        for option in list(section):
             option_name, option_value = _read_option(section_name, option)
             if type(option_value) is list:
                 for ov in option_value:
@@ -175,7 +178,7 @@ def _ossecconf2json(xml_conf):
     """
     final_json = {}
 
-    for root in xml_conf.getchildren():
+    for root in list(xml_conf):
         if root.tag.lower() == "ossec_config":
             _conf2json(root, final_json)
 
@@ -420,7 +423,8 @@ def get_agent_conf(group_id=None, offset=0, limit=common.database_limit, filenam
     """
 
     if group_id:
-        if not Agent.group_exists(group_id):
+        # if not group_exists(group_id):
+        if not os_path.exists("{0}/{1}".format(common.shared_path, group_id)):
             raise WazuhException(1710, group_id)
 
         agent_conf = "{0}/{1}".format(common.shared_path, group_id)
@@ -463,7 +467,7 @@ def get_file_conf(filename, group_id=None, type_conf=None):
     """
 
     if group_id:
-        if not Agent.group_exists(group_id):
+        if not group_exists(group_id):
             raise WazuhException(1710, group_id)
 
         file_path = "{0}/{1}".format(common.shared_path, filename) \