+++ framework/wazuh/manager.py
@@ -3,15 +3,18 @@
 # Created by Wazuh, Inc. <info@wazuh.com>.
 # This program is a free software; you can redistribute it and/or modify it under the terms of GPLv2
 
+from wazuh.exception import WazuhException
 from wazuh.utils import execute, previous_month, cut_array, sort_array, search_array, tail
 from wazuh import common
 from datetime import datetime
+from wazuh.configuration import get_ossec_conf
+from wazuh.cluster.distributed_api import is_a_local_request, distributed_api_request, is_cluster_running
+from wazuh.cluster.protocol_messages import list_requests_managers
 import time
 from os.path import exists
 from glob import glob
 import re
 import hashlib
-
 
 def status():
     """
@@ -19,9 +22,9 @@ def status():
     :return: Array of dictionaries (keys: status, daemon).
     """
 
-    processes = ['ossec-monitord', 'ossec-logcollector', 'ossec-remoted', 
-                 'ossec-syscheckd', 'ossec-analysisd', 'ossec-maild', 
-                 'ossec-execd', 'wazuh-modulesd', 'ossec-authd', 
+    processes = ['ossec-monitord', 'ossec-logcollector', 'ossec-remoted',
+                 'ossec-syscheckd', 'ossec-analysisd', 'ossec-maild',
+                 'ossec-execd', 'wazuh-modulesd', 'ossec-authd',
                  'wazuh-clusterd']
 
     data = {}
@@ -160,3 +163,50 @@ def ossec_log_summary(months=3):
             else:
                 continue
     return categories
+
+
+def managers_status(node_id=None, cluster_depth=1):
+    if is_a_local_request() or cluster_depth <= 0 :
+        return status()
+    else:
+        if not is_cluster_running():
+            raise WazuhException(3015)
+
+        request_type = list_requests_managers['MANAGERS_STATUS']
+        return distributed_api_request(request_type=request_type, cluster_depth=cluster_depth, affected_nodes=node_id)
+
+
+def managers_ossec_log(type_log='all', category='all', months=3, offset=0, limit=common.database_limit, sort=None, search=None, node_id=None, cluster_depth=1):
+    if is_a_local_request() or cluster_depth <= 0 :
+        return ossec_log(type_log, category, months, offset, limit, sort, search)
+    else:
+        if not is_cluster_running():
+            raise WazuhException(3015)
+
+        request_type = list_requests_managers['MANAGERS_LOGS']
+        args = [str(type_log), str(category), str(months), str(offset), str(limit), str(sort), str(search)]
+        return distributed_api_request(request_type=request_type, args=args, cluster_depth=cluster_depth, affected_nodes=node_id)
+
+
+def managers_ossec_log_summary(months=3, node_id=None, cluster_depth=1):
+    if is_a_local_request() or cluster_depth <= 0 :
+        return ossec_log_summary(months)
+    else:
+        if not is_cluster_running():
+            raise WazuhException(3015)
+
+        request_type = list_requests_managers['MANAGERS_LOGS_SUMMARY']
+        args = [str(months)]
+        return distributed_api_request(request_type=request_type, args=args, cluster_depth=cluster_depth, affected_nodes=node_id)
+
+
+def managers_get_ossec_conf(section=None, field=None, node_id=None, cluster_depth=1):
+    if is_a_local_request() or cluster_depth <= 0 :
+        return get_ossec_conf(section, field)
+    else:
+        if not is_cluster_running():
+            raise WazuhException(3015)
+
+        request_type = list_requests_managers['MANAGERS_OSSEC_CONF']
+        args = [str(section), str(field)]
+        return distributed_api_request(request_type=request_type, args=args, cluster_depth=cluster_depth, affected_nodes=node_id)