+++ framework/wazuh/rootcheck.py
@@ -11,17 +11,11 @@ from wazuh.ossec_queue import OssecQueue
 from wazuh import common
 from glob import glob
 from os import remove, path
+from wazuh.cluster.distributed_api import is_a_local_request, distributed_api_request, is_cluster_running
+from wazuh.cluster.protocol_messages import list_requests_rootcheck
 
 
-def run(agent_id=None, all_agents=False):
-    """
-    Runs rootcheck and syscheck.
-
-    :param agent_id: Run rootcheck/syscheck in the agent.
-    :param all_agents: Run rootcheck/syscheck in all agents.
-    :return: Message.
-    """
-
+def run_local(agent_id=None, all_agents=False):
     if agent_id == "000" or all_agents:
         try:
             SYSCHECK_RESTART = "{0}/var/run/.syscheck_run".format(common.ossec_path)
@@ -55,15 +49,26 @@ def run(agent_id=None, all_agents=False):
     return ret_msg
 
 
-def clear(agent_id=None, all_agents=False):
+def run(agent_id=None, all_agents=False, cluster_depth=1):
     """
-    Clears the database.
+    Runs rootcheck and syscheck.
 
-    :param agent_id: For an agent.
-    :param all_agents: For all agents.
+    :param agent_id: Run rootcheck/syscheck in the agent.
+    :param all_agents: Run rootcheck/syscheck in all agents.
     :return: Message.
     """
+    if is_a_local_request() or agent_id == "000" or cluster_depth <= 0:
+        return run_local(agent_id, all_agents)
+    else:
+        if not is_cluster_running():
+            raise WazuhException(3015)
 
+        request_type = list_requests_rootcheck['ROOTCHECK_RUN']
+        args = [str(all_agents)]
+        return distributed_api_request(request_type, Agent.get_agents_by_node(agent_id), args, cluster_depth)
+
+
+def clear_local(agent_id=None, all_agents=False):
     # Clear DB
     if int(all_agents):
         db_agents = glob('{0}/*-*.db'.format(common.database_path_agents))
@@ -99,6 +104,25 @@ def clear(agent_id=None, all_agents=False):
             remove(rootcheck_file)
 
     return "Rootcheck database deleted"
+
+
+def clear(agent_id=None, all_agents=False, cluster_depth=1):
+    """
+    Clears the database.
+
+    :param agent_id: For an agent.
+    :param all_agents: For all agents.
+    :return: Message.
+    """
+    if is_a_local_request() or agent_id == "000" or cluster_depth <= 0:
+        return clear_local(agent_id, all_agents)
+    else:
+        if not is_cluster_running():
+            raise WazuhException(3015)
+
+        request_type = list_requests_rootcheck['ROOTCHECK_CLEAR']
+        args = [str(all_agents)]
+        return distributed_api_request(request_type, Agent.get_agents_by_node(agent_id), args, cluster_depth)
 
 
 def print_db(agent_id=None, status='all', pci=None, cis=None, offset=0, limit=common.database_limit, sort=None, search=None):
@@ -168,7 +192,7 @@ def print_db(agent_id=None, status='all', pci=None, cis=None, offset=0, limit=co
             if not set(sort['fields']).issubset(allowed_sort_fields):
                 uncorrect_fields = map(lambda x: str(x), set(sort['fields']) - set(allowed_sort_fields))
                 raise WazuhException(1403, 'Allowed sort fields: {0}. Fields: {1}'.format(allowed_sort_fields, uncorrect_fields))
-                
+
             query += ' ORDER BY ' + ','.join(['{0} {1}'.format(fields[i], sort['order']) for i in sort['fields']])
         else:
             query += ' ORDER BY date_last {0}'.format(sort['order'])
@@ -206,18 +230,7 @@ def print_db(agent_id=None, status='all', pci=None, cis=None, offset=0, limit=co
     return data
 
 
-def get_pci(agent_id=None, offset=0, limit=common.database_limit, sort=None, search=None):
-    """
-    Get all the PCI requirements used in the rootchecks of the agent.
-
-    :param agent_id: Agent ID.
-    :param offset: First item to return.
-    :param limit: Maximum number of items to return.
-    :param sort: Sorts the items. Format: {"fields":["field1","field2"],"order":"asc|desc"}.
-    :param search: Looks for items with the specified string.
-    :return: Dictionary: {'items': array of items, 'totalItems': Number of items (without applying the limit)}
-    """
-
+def get_pci_local(agent_id=None, offset=0, limit=common.database_limit, sort=None, search=None):
     query = "SELECT {0} FROM pm_event WHERE pci_dss IS NOT NULL"
     fields = {}
     request = {}
@@ -271,9 +284,9 @@ def get_pci(agent_id=None, offset=0, limit=common.database_limit, sort=None, sea
     return data
 
 
-def get_cis(agent_id=None, offset=0, limit=common.database_limit, sort=None, search=None):
+def get_pci(agent_id=None, offset=0, limit=common.database_limit, sort=None, search=None):
     """
-    Get all the CIS requirements used in the rootchecks of the agent.
+    Get all the PCI requirements used in the rootchecks of the agent.
 
     :param agent_id: Agent ID.
     :param offset: First item to return.
@@ -282,7 +295,18 @@ def get_cis(agent_id=None, offset=0, limit=common.database_limit, sort=None, sea
     :param search: Looks for items with the specified string.
     :return: Dictionary: {'items': array of items, 'totalItems': Number of items (without applying the limit)}
     """
+    if is_a_local_request() or agent_id == "000":
+        return get_pci_local(agent_id, offset, limit, sort, search)
+    else:
+        if not is_cluster_running():
+            raise WazuhException(3015)
 
+        request_type = list_requests_rootcheck['ROOTCHECK_PCI']
+        args = [str(offset), str(limit), str(sort), str(search)]
+        return distributed_api_request(request_type, Agent.get_agents_by_node(agent_id), args)
+
+
+def get_cis_local(agent_id=None, offset=0, limit=common.database_limit, sort=None, search=None):
     query = "SELECT {0} FROM pm_event WHERE cis IS NOT NULL"
     fields = {}
     request = {}
@@ -336,13 +360,29 @@ def get_cis(agent_id=None, offset=0, limit=common.database_limit, sort=None, sea
     return data
 
 
-def last_scan(agent_id):
+def get_cis(agent_id=None, offset=0, limit=common.database_limit, sort=None, search=None):
     """
-    Gets the last scan of the agent.
+    Get all the CIS requirements used in the rootchecks of the agent.
 
     :param agent_id: Agent ID.
-    :return: Dictionary: end, start.
+    :param offset: First item to return.
+    :param limit: Maximum number of items to return.
+    :param sort: Sorts the items. Format: {"fields":["field1","field2"],"order":"asc|desc"}.
+    :param search: Looks for items with the specified string.
+    :return: Dictionary: {'items': array of items, 'totalItems': Number of items (without applying the limit)}
     """
+    if is_a_local_request() or agent_id == "000":
+        return get_cis_local(agent_id, offset, limit, sort, search)
+    else:
+        if not is_cluster_running():
+            raise WazuhException(3015)
+
+        request_type = list_requests_rootcheck['ROOTCHECK_CIS']
+        args = [str(offset), str(limit), str(sort), str(search)]
+        return distributed_api_request(request_type, Agent.get_agents_by_node(agent_id), args)
+
+
+def last_scan_local(agent_id):
     # Connection
     db_agent = glob('{0}/{1}-*.db'.format(common.database_path_agents, agent_id))
     if not db_agent:
@@ -366,3 +406,21 @@ def last_scan(agent_id):
         data['start'] = tuple[0]
 
     return data
+
+
+def last_scan(agent_id):
+    """
+    Gets the last scan of the agent.
+
+    :param agent_id: Agent ID.
+    :return: Dictionary: end, start.
+    """
+    if is_a_local_request() or agent_id == "000":
+        return last_scan_local(agent_id)
+    else:
+        if not is_cluster_running():
+            raise WazuhException(3015)
+
+        request_type = list_requests_rootcheck['ROOTCHECK_LAST_SCAN']
+        args = []
+        return distributed_api_request(request_type, Agent.get_agents_by_node(agent_id), args)