+++ src/wazuh_modules/wmodules.h
@@ -22,6 +22,7 @@
 
 #define WM_DEFAULT_DIR  DEFAULTDIR "/wodles"        // Default modules directory.
 #define WM_STATE_DIR    DEFAULTDIR "/var/wodles"    // Default directory for states.
+#define WM_DIR_WIN      "wodles"                    // Default directory for states (Windows)
 #define WM_STRING_MAX   67108864                    // Max. dynamic string size (64 MB).
 #define WM_BUFFER_MAX   1024                        // Max. static buffer size.
 #define WM_MAX_ATTEMPTS 3                           // Max. number of attempts.
@@ -61,6 +62,9 @@ extern wmodule *wmodules;       // Loaded modules.
 extern int wm_task_nice;        // Nice value for tasks.
 extern int wm_max_eps;          // Maximum events per second sent by OpenScap Wazuh Module
 extern int wm_kill_timeout;     // Time for a process to quit before killing it
+
+// Read XML configuration and internal options
+int wm_config();
 
 // Add module to the global list
 void wm_add(wmodule *module);