+++ CHANGELOG.md
@@ -4,13 +4,27 @@ All notable changes to this project will be documented in this file.
 ## [v3.2.0]
 
 ### Added
-
+- [Added support for multiple masters in cluster.](https://github.com/wazuh/wazuh/pull/353)
+- [Added type of node in API call `GET/cluster/nodes` and `cluster_control --nodes`](https://github.com/wazuh/wazuh/pull/353)
+- [Added support to synchronize custom rules and decoders in the cluster.](https://github.com/wazuh/wazuh/pull/344)
 - [Add field `status` to `GET/agents/groups/:group_id` API call.](https://github.com/wazuh/wazuh/pull/338)
+- [Added support to synchronize custom rules and decoders in the cluster.](https://github.com/wazuh/wazuh/pull/344)
+- [Add field `status` to `GET/agents/groups/:group_id` API call.](https://github.com/wazuh/wazuh/pull/338)
+- [Added support for multiple masters in cluster.](https://github.com/wazuh/wazuh/pull/353)
+- [Added type of node in API call `GET/cluster/nodes` and `cluster_control --nodes`](https://github.com/wazuh/wazuh/pull/353)
+- Added support for API in cluster ([#362](https://github.com/wazuh/wazuh/pull/362)).
+- Added support for Windows to CIS-CAT integration module ([#369](https://github.com/wazuh/wazuh/pull/369))
+
+### Changed
+- [Return API call `GET/cluster/files` by node name instead of by node IP.](https://github.com/wazuh/wazuh/pull/353)
 
 ### Fixed
-
 - [Fixed oscap.py to support new versions of OpenSCAP scanner.](https://github.com/wazuh/wazuh/pull/331)
+- [Fixed timeout bug when the cluster port was closed.](https://github.com/wazuh/wazuh/pull/343)
+- [Improve exception handling in `cluster_control`.](https://github.com/wazuh/wazuh/pull/343)
 - [Fixed bug in cluster when receive an error response from client.](https://github.com/wazuh/wazuh/pull/346)
+- [Improved cluster logs](https://github.com/wazuh/wazuh/pull/353)
+- [Fixed bug in framework when the manager is installed in different path than /var/ossec.](https://github.com/wazuh/wazuh/pull/335)
 
 ## [v3.1.0]